ml_part/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-generativeai>=0.1.0rc2
pandas>=1.5.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
python-multipart>=0.0.6
